+++ src/views/app/detail-panel/edit/editor/tests/edit-view.test.tsx
@@ -11,7 +11,6 @@ import React, { useState } from 'react';
 
 import { faker } from '@faker-js/faker';
 import { act, waitFor, fireEvent } from '@testing-library/react';
-import { UserEvent } from '@testing-library/user-event';
 import * as hooks from '@zextras/carbonio-shell-ui';
 import { ErrorSoapBodyResponse } from '@zextras/carbonio-shell-ui';
 import { ParticipantRole } from '@zextras/carbonio-ui-commons';
@@ -22,7 +21,7 @@ import type { Mock } from 'vitest';
 import { aSuccessfulSaveDraft, aFailingSaveDraft } from './utils/utils';
 import { TESTID_SELECTORS } from '../../../../../../__test__/constants';
 import { EditView } from '../edit-view';
-import { setupTest, screen, within } from '@test-setup';
+import { setupTest, screen, within, UserEvent } from '@test-setup';
 import {
 	createSoapAPIInterceptor,
 	createAPIInterceptor
@@ -163,7 +162,7 @@ const getSubjectInput = (): HTMLElement =>
 const getEditorTextareaElement = (): HTMLInputElement => screen.getByTestId('MailPlainTextEditor');
 
 const makeSomeChangeToTriggerSaveDraft = async (user: UserEvent): Promise<void> => {
-	await user.type(getSubjectInput(), 'Some subject');
+	await user.pasteInto(getSubjectInput(), 'Some subject');
 };
 
 describe('Edit view', () => {
@@ -344,7 +343,7 @@ describe('Edit view', () => {
 
 			// Insert a subject
 			await act(async () => {
-				await user.type(getSubjectInput(), subject);
+				await user.pasteInto(getSubjectInput(), subject);
 			});
 
 			const optionIcon = screen.getByTestId('options-dropdown-icon');